# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hodler Scanner API",
    "description": "Buy & Hold technical analysis — SMA, RSI, MACD, Bollinger Bands, scoring",
    "version": "1.1.1-SNAPSHOT"
  },
  "paths": {
...
    "/search": {
      "get": {
        "tags": [
          "utils"
        ],
        "summary": "Search tickers by name or symbol",
        "description": "Search Yahoo Finance by company name, ticker symbol or **ISIN**\n(e.g. `Apple`, `MC.PA`, `FR0000121014`). Fuzzy matching tolerates typos.",
        "operationId": "search_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response 200 Search Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{ticker_code}": {
      "get": {
        "summary": "Full technical analysis of a ticker",
        "description": "Return all technical indicators + B&H score for a ticker.\n\n**Examples**: `MC.PA`, `AAPL`, `ASML.AS`, `BTC-USD`\n\n- `refresh=true`: force recomputation even if the cache is valid\n- `lang`: language of the analysis text (`en` by default, `fr`)",
        "operationId": "get_ticker_ticker__ticker_code__get",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "en",
              "title": "Lang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Analysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{ticker_code}/chart": {
      "get": {
        "tags": [
          "analyse"
        ],
        "summary": "Historical data for charts",
        "description": "Return price, SMA 200/50, RSI 14 and MACD histogram for the charts.",
        "operationId": "ticker_chart_ticker__ticker_code__chart_get",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1y",
              "title": "Period"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{ticker_code}/fundamentals": {
      "get": {
        "tags": [
          "analyse"
        ],
        "summary": "Fundamental data",
        "description": "P/E, market cap, sector, country via yfinance.info.",
        "operationId": "ticker_fundamentals_ticker__ticker_code__fundamentals_get",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fundamentals"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{ticker_code}/news": {
      "get": {
        "tags": [
          "analyse"
        ],
        "summary": "Recent news",
        "description": "Latest news for the stock via yfinance (Yahoo Finance).",
        "operationId": "ticker_news_ticker__ticker_code__news_get",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/News"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{ticker_code}/backtest": {
      "get": {
        "tags": [
          "analyse"
        ],
        "summary": "Light score backtest",
        "description": "Replay the Buy & Hold score over ~5 years of history and report the realized\nforward returns (≈3M / 6M / 12M) grouped by score band — a credibility check\non the scoring engine.",
        "operationId": "ticker_backtest_ticker__ticker_code__backtest_get",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/cache/{ticker_code}": {
      "delete": {
        "tags": [
          "cache"
        ],
        "summary": "Invalidate a ticker",
        "operationId": "cache_invalidate_cache__ticker_code__delete",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "tags": [
          "cache"
        ],
        "summary": "Clear selected caches and DB data",
        "description": "Selectively wipe in-memory caches and DB rows per data type.\n\nEach flag is opt-in: ``caches`` (in-memory analyses/charts/news…),\n``watchlist`` (favorites), ``portfolio`` (positions), ``backtests``\n(stored score history) and ``tickers`` (the name/ISIN dictionary).",
        "operationId": "reset_all_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/favorites": {
      "get": {
        "tags": [
          "favorites"
        ],
        "summary": "List favorites",
        "description": "Return the watchlist persisted server-side.",
        "operationId": "favorites_list_favorites_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "favorites"
        ],
        "summary": "Replace the entire favorites list",
        "description": "Fully replace the watchlist (used for import/migration).",
        "operationId": "favorites_set_favorites_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoritesImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/favorites/{ticker_code}": {
      "post": {
        "tags": [
          "favorites"
        ],
        "summary": "Add a favorite",
        "operationId": "favorites_add_favorites__ticker_code__post",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "favorites"
        ],
        "summary": "Remove a favorite",
        "operationId": "favorites_remove_favorites__ticker_code__delete",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoritesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/portfolio/{ticker_code}": {
      "put": {
        "tags": [
          "portfolio"
        ],
        "summary": "Add or update a position",
        "description": "Insert or update a position (one aggregated lot per ticker).",
        "operationId": "portfolio_upsert_portfolio__ticker_code__put",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Remove a position",
        "operationId": "portfolio_remove_portfolio__ticker_code__delete",
        "parameters": [
          {
            "name": "ticker_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Portfolio"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tickers": {
      "post": {
        "summary": "Analyze a list of tickers",
        "description": "Analyze several tickers in parallel.\n\n```json\n{ \"tickers\": [\"MC.PA\", \"ASML.AS\", \"AIR.PA\"], \"refresh\": false, \"lang\": \"en\" }\n```\n\nEach entry in the result contains either the data or an `error` field.",
        "operationId": "post_tickers_tickers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickerListRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Analysis": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "data_partiel": {
            "type": "boolean",
            "title": "Data Partiel"
          },
          "days_available": {
            "type": "integer",
            "title": "Days Available"
          },
          "price": {
            "$ref": "#/components/schemas/AnalysisPrice"
          },
          "indicators": {
            "$ref": "#/components/schemas/AnalysisIndicators"
          },
          "distances": {
            "$ref": "#/components/schemas/AnalysisDistances"
          },
          "fundamentals": {
            "$ref": "#/components/schemas/AnalysisFundamentals"
          },
          "signals": {
            "$ref": "#/components/schemas/AnalysisSignals"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisBlock"
          },
          "cached": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "timestamp",
          "data_partiel",
          "days_available",
          "price",
          "indicators",
          "distances",
          "fundamentals",
          "signals",
          "analysis"
        ],
        "title": "Analysis"
      },
...
      "FavoritesImport": {
        "properties": {
          "tickers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tickers"
          }
        },
        "type": "object",
        "required": [
          "tickers"
        ],
        "title": "FavoritesImport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchResult": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "exchange": {
            "type": "string",
            "title": "Exchange"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name",
          "type",
          "exchange"
        ],
        "title": "SearchResult"
      },
...
    }
  }
}